make Candidates communicate only with Groups
Each Candidate sends its result to its Groups and does not communicate
with adjacent Candidates. Each Group tracks the information received
from its Candidates and sends one value to all of its Candidates under
whichever of the following conditions occurs first: if a Candidate
reports that it is true, the Group sends false to all of its
Candidates; or, if all Candidates but one report that they are false,
it sends true to all Candidates. Each Candidate acts only on the first
value it receives and ignores subsequent values. (In a well-formed
sudoku puzzle with a single solution, this is not a race condition
because the truth value of a Candidate inference does not depend on the
order of the logical deductions that prove it.)

Therefore, each Candidate receives at most five values on its channel,
and each Group precisely nine on its. Candidates do not need to send to
each other, so they do not require knowledge of each other's adjacency.

remove ach

remove isAdjacent

clarify doc comments

declare nGroups globally

factor out sendAll

rename cch, gch, and b fields to cans, groups, and value